// src/app/modules/character/models/character.enum.ts
export enum CharactersActions {
    SetCharacterSortOption = '[Character List Page] Set Characters Sort Option',
    SearchCharacters = '[Character List Page] Search Characters',
    LoadMoreCharacters = '[Character List Page] Load More Characters',
    LoadCharacters = '[Character List Page] Load Characters',
    CharactersLoadedSuccess = '[Character List Page] Characters Loaded Success',
    CharactersLoadedError = '[Character List Page] Characters Loaded Error',
}